+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `tango` command.
+
 ## [v0.0.2](https://github.com/allenai/tango/releases/tag/v0.0.2) - 2021-09-27
 
 ### Added